requirements.txt
Django==1.8.5
django-cors-headers==1.1.0
djangorestframework==3.3.1
Pillow==3.0.0
psycopg2==2.6.1
wheel==0.24.0
